test: split query settings e2e specs
